+++ src/main/java/com/facilio/bmsconsole/commands/AddOrUpdateWorkorderServiceCommand.java
@@ -114,7 +114,7 @@ public class AddOrUpdateWorkorderServiceCommand extends FacilioCommand{
 				duration = getEstimatedWorkDuration(woService.getStartTime(), woService.getEndTime());
 			} else {
 				if(workorder.getActualWorkDuration() > 0) {
-					double hours = (workorder.getActualWorkDuration() / (60 * 60));
+					double hours = (((double)workorder.getActualWorkDuration()) / (60 * 60));
 					duration = Math.round(hours*100.0)/100.0;
 				}
 				else{
